Started work on friends system
Currently we have FriendshipRequests, which you can send to a user. You
can see incoming requests, and who you've sent requests to. We are
planning on having a bijective graph of users, which is possible but
we're not quite sure how to represent that in the ORM.

Next task is going to be FriendAccept implementation.