Implemented FriendAccept and FriendRemove
We also forgot to add a FriendList, so that is in there too. At the
moment friendship is non directional, and is represented by two rows in
the friendship table. One says a->b and the other b->a.

They both need to be created so please use the
Friendship.{add_friendship,remove_friendship,add_pair,remove_pair)
static functions if you want to modify a friendship.