ADD image upload and retrieval
At the moment, no authoriziation is performed on upload/download, just
authentication. The image is resized to a maximum of 800x600 and then
saved with a random filename in the configured directory.

It might be hard to install graphicsmagick on other computers, but let
me know how it goes. I might write a quick check and failure if it isn't
installed